Exercise 2.1

+++ src/guessing-game/game.cpp
@@ -8,8 +8,8 @@ int main()
     std::rand(); 
     static const double base = 1/(RAND_MAX + 1);
     std::rand(); 
-    static const uint8_t target_number = 1 + static_cast<uint8_t>(100*std::rand()*base);
-    uint8_t user_guess = 0; 
+    static const int target_number = 1 + static_cast<int>(100*std::rand()*base);
+    int user_guess = 0; 
     std::cout << " Guess a number between 1 - 100 \n"; 
     std::cin >> user_guess; 
     while(user_guess != target_number)
